refactor: updated tests
moved fixtures to `conftest.py` and enforced styling
